Added Character Dialogue & NPC
There is now an NPC in the scene. I added the colliders and everything to the NPC, made sure the tags were set, and had the dialogue box in the flowchart set. It's still not working when I walk into the collider. Not sure why, maybe someone else could take a look at it. Also the clicking on the ground also doesn't work on the colliders.