Fixed Character Issues & Added Flower Quest
The player had an issue where half the character model was below the ground. It is now above the ground where it should be and isn't rotating uncontrollably now because the rigidbody is not conflicting with the ground. Also I added in the flower quest, where the player talks to the Jack NPC and sets the player out to retrieve a flower and bring it back to him.