Rearranged Map
The environment is set up so it makes sense to the player. There are paths that are blocked by the environment and trees & such. Along those paths, the player is able to see other players/quests/landmarks whilst on their journey from the zone of influence. I've also removed the 4th option (ignore option) in the dialogue, so there are only 3 options now.